# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/guides": {
      "get": {
        "description": "Returns list of all guides. Accessible by superuser, admin, traveler, guide, and agency.",
        "operationId": "GuidesController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of guides returned successfully"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "summary": "Get all guides",
        "tags": [
          "guides"
        ]
      },
      "post": {
        "description": "Creates a guide account. Accessible by superuser and admin.",
        "operationId": "GuidesController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guide created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "summary": "Create a new guide",
        "tags": [
          "guides"
        ]
      }
    },
    "/guides/{id}": {
      "get": {
        "description": "Returns a single guide. Accessible by superuser, admin, traveler, guide, and agency.",
        "operationId": "GuidesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Guide ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guide found"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Guide not found"
          }
        },
        "summary": "Get guide by ID",
        "tags": [
          "guides"
        ]
      },
      "put": {
        "description": "Updates guide details. Accessible by superuser, admin, and the guide themselves.",
        "operationId": "GuidesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Guide ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Guide updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Guide not found"
          }
        },
        "summary": "Update a guide",
        "tags": [
          "guides"
        ]
      },
      "delete": {
        "description": "Removes a guide. Accessible by superuser and admin.",
        "operationId": "GuidesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Guide ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guide deleted successfully"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Guide not found"
          }
        },
        "summary": "Delete a guide",
        "tags": [
          "guides"
        ]
      }
    },
...
    "/trips/guide/{guideId}": {
      "get": {
        "description": "Returns all trips assigned to a guide.",
        "operationId": "TripsController_findByGuide",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trips for guide"
          }
        },
        "summary": "Get trips by guide ID",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/itineraries/trip/{tripId}": {
      "get": {
        "description": "Returns all itinerary items for a trip.",
        "operationId": "ItinerariesController_findByTrip",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Itinerary items for trip"
          }
        },
        "summary": "Get itinerary by trip ID",
        "tags": [
          "itineraries"
        ]
      }
    },
    "/itineraries": {
      "post": {
        "operationId": "ItinerariesController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItineraryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Itinerary created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create itinerary item",
        "tags": [
          "itineraries"
        ]
      }
    },
    "/itineraries/{id}": {
      "put": {
        "operationId": "ItinerariesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItineraryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Itinerary updated"
          },
          "404": {
            "description": "Itinerary not found"
          }
        },
        "summary": "Update itinerary item",
        "tags": [
          "itineraries"
        ]
      },
      "delete": {
        "operationId": "ItinerariesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Itinerary deleted"
          },
          "404": {
            "description": "Itinerary not found"
          }
        },
        "summary": "Delete itinerary item",
        "tags": [
          "itineraries"
        ]
      }
    },
...
    "/dashboard/guide/{guideId}": {
      "get": {
        "description": "Assignment metrics: assigned trips, completed, pending, messages.",
        "operationId": "DashboardController_getGuideDashboard",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guide dashboard data"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Guide dashboard",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "GlobeSync API",
    "description": "GlobeSync Travel Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateGuideDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sophia",
            "description": "Guide name"
          },
          "email": {
            "type": "string",
            "example": "sophia@gmail.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "guide123",
            "description": "Account password"
          },
          "languages": {
            "example": [
              "English",
              "Spanish"
            ],
            "description": "Languages spoken",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "experience": {
            "type": "number",
            "example": 5,
            "description": "Years of experience"
          },
          "rating": {
            "type": "number",
            "example": 4.5,
            "description": "Rating (0-5)"
          },
          "pricePerTrip": {
            "type": "number",
            "example": 7000,
            "description": "Price per trip"
          },
          "contact": {
            "type": "string",
            "example": "9876543210",
            "description": "Contact number"
          },
          "adminId": {
            "type": "number",
            "example": 1,
            "description": "Assigned admin ID"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UpdateGuideDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sophia",
            "description": "Guide name"
          },
          "email": {
            "type": "string",
            "example": "sophia@gmail.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "newpassword",
            "description": "New password"
          },
          "languages": {
            "example": [
              "English",
              "Spanish"
            ],
            "description": "Languages spoken",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "experience": {
            "type": "number",
            "example": 5,
            "description": "Years of experience"
          },
          "rating": {
            "type": "number",
            "example": 4.5,
            "description": "Rating (0-5)"
          },
          "pricePerTrip": {
            "type": "number",
            "example": 7000,
            "description": "Price per trip"
          },
          "contact": {
            "type": "string",
            "example": "9876543210",
            "description": "Contact number"
          },
          "adminId": {
            "type": "number",
            "example": 1,
            "description": "Assigned admin ID"
          }
        }
      },
...
      "CreateItineraryDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "number",
            "example": 1
          },
          "city": {
            "type": "string",
            "example": "Paris"
          },
          "dayNumber": {
            "type": "number",
            "example": 1
          },
          "activity": {
            "type": "string",
            "example": "Visit Eiffel Tower"
          },
          "activityStatus": {
            "type": "string",
            "example": "Pending",
            "enum": [
              "Pending",
              "Confirmed",
              "Completed"
            ]
          },
          "time": {
            "type": "string",
            "example": "10:00"
          }
        },
        "required": [
          "tripId",
          "city",
          "dayNumber",
          "activity",
          "time"
        ]
      },
      "UpdateItineraryDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "number",
            "example": 1
          },
          "city": {
            "type": "string",
            "example": "Paris"
          },
          "dayNumber": {
            "type": "number",
            "example": 1
          },
          "activity": {
            "type": "string",
            "example": "Visit Eiffel Tower"
          },
          "activityStatus": {
            "type": "string",
            "example": "Confirmed",
            "enum": [
              "Pending",
              "Confirmed",
              "Completed"
            ]
          },
          "time": {
            "type": "string",
            "example": "10:00"
          }
        }
      },
...
    }
  }
}